refactor agent api to mvc design
