+++ src/api/server.ts
@@ -1,17 +1,16 @@
-import express, { Response } from 'express';
+import express from 'express';
 import cors from 'cors';
 import { createLogger } from '../utils/logger.js';
 import { OrchestratorRunner } from '../agents/workflows/orchestrator/orchestratorWorkflow.js';
 import { ApiServer, LogMetadata } from './types.js';
 import { config } from '../config/index.js';
 import { createApiRouter } from './routes/index.js';
-
+import { broadcastTaskUpdateUtility } from './controller/TaskController.js';
 import { Logger } from 'winston';
+import { attachLoggerUtility } from './controller/LogsController.js';
+import { broadcastLogUtility } from './controller/LogsController.js';
+import { getRegisteredNamespaces, registerRunnerUtility } from './controller/WorkflowController.js';
 const logger = createLogger('api-server');
-
-export const logStreamClients = new Map<number, { res: Response; namespace: string }>();
-export const taskStreamClients = new Map<number, { res: Response; namespace: string }>();
-export const orchestratorRunners = new Map<string, OrchestratorRunner>();
 
 let apiServer: ApiServer | null = null;
 
@@ -28,88 +27,14 @@ const createSingletonApiServer = (): ApiServer => {
     logger.info(`API server started on port ${PORT}`);
   });
 
-  const broadcastLog = (namespace: string, level: string, message: string, meta?: LogMetadata) => {
-    logStreamClients.forEach((client, clientId) => {
-      if (client.namespace === namespace || client.namespace === 'all') {
-        const logEvent = {
-          type: 'log',
-          timestamp: new Date().toISOString(),
-          namespace,
-          level,
-          message,
-          meta,
-        };
-
-        try {
-          client.res.write(`data: ${JSON.stringify(logEvent)}\n\n`);
-        } catch (e) {
-          logger.error('Error sending to client, removing client', { clientId, error: e });
-          logStreamClients.delete(clientId);
-        }
-      }
-    });
-  };
-
-  const broadcastTaskUpdate = (namespace: string) => {
-    taskStreamClients.forEach((client, clientId) => {
-      if (client.namespace === namespace || client.namespace === 'all') {
-        const runner = orchestratorRunners.get(namespace);
-        if (!runner || typeof runner.getTaskQueue !== 'function') {
-          return;
-        }
-
-        const tasks = runner.getTaskQueue();
-        const taskEvent = {
-          type: 'tasks',
-          timestamp: new Date().toISOString(),
-          namespace,
-          tasks,
-        };
-
-        try {
-          client.res.write(`data: ${JSON.stringify(taskEvent)}\n\n`);
-        } catch (e) {
-          logger.error('Error sending task update to client, removing client', {
-            clientId,
-            error: e,
-          });
-          taskStreamClients.delete(clientId);
-        }
-      }
-    });
-  };
-
-  const attachLogger = (existingLogger: Logger, namespace: string): Logger => {
-    type LoggerMethod = (message: string | object, ...meta: unknown[]) => Logger;
-    const wrapLoggerMethod = (level: string, method: LoggerMethod): LoggerMethod => {
-      return function (this: Logger, message: string | object, ...meta: unknown[]): Logger {
-        const result = method.apply(this, [message, ...meta]);
-        broadcastLog(namespace, level, String(message), meta[0] as LogMetadata);
-        return result;
-      };
-    };
-
-    existingLogger.info = wrapLoggerMethod('info', existingLogger.info);
-    existingLogger.debug = wrapLoggerMethod('debug', existingLogger.debug);
-    existingLogger.error = wrapLoggerMethod('error', existingLogger.error);
-
-    return existingLogger;
-  };
-
   return {
     app,
     server,
-    registerRunner: (namespace: string, runner: OrchestratorRunner) => {
-      orchestratorRunners.set(namespace, runner);
-      logger.info(`Registered orchestrator runner with namespace: ${namespace}`);
-      return { namespace, runner };
-    },
-    broadcastLog,
-    broadcastTaskUpdate,
-    attachLogger,
-    getRegisteredNamespaces: () => {
-      return Array.from(orchestratorRunners.keys());
-    },
+    registerRunner: registerRunnerUtility,
+    broadcastLog: broadcastLogUtility,
+    broadcastTaskUpdate: broadcastTaskUpdateUtility,
+    attachLogger: attachLoggerUtility,
+    getRegisteredNamespaces: getRegisteredNamespaces,
   };
 };
